refactor(derive): Move from_arg_matches into coupled derives
