+++ clap_derive/src/derives/mod.rs
@@ -12,8 +12,8 @@
 // commit#ea76fa1b1b273e65e3b0b1046643715b49bec51f which is licensed under the
 // MIT/Apache 2.0 license.
 mod arg_enum;
+mod args;
 mod clap;
-mod from_arg_matches;
 mod into_app;
 mod subcommand;
